# services/logger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AIctivity API",
    "description": "This API is used to save activity data.",
    "license": {
      "name": "TODO",
      "url": "http://todo.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:6000",
      "variables": {}
    }
  ],
  "paths": {
    "/data": {
      "post": {
        "summary": "Save new Data",
        "description": "Save new accelerometer, gyroscope and temp data",
        "operationId": "saveData",
        "parameters": [],
        "requestBody": {
          "description": "Save Data",
          "content": {
            "application/json": {
              "schema": {
                "description": "Data to Save",
                "$ref": "#/components/schemas/Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Data Saved",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Bad Request Error",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "deprecated": false,
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Data": {
        "title": "Data",
        "type": "object",
        "properties": {
          "aX": {
            "type": "integer"
          },
          "aY": {
            "type": "integer"
          },
          "aZ": {
            "type": "integer"
          },
          "gX": {
            "type": "integer"
          },
          "gY": {
            "type": "integer"
          },
          "gZ": {
            "type": "integer"
          },
          "temp": {
            "type": "number"
          } 
        }
      },"Error": {
        "title": "Error",
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  } 
}